// packages/client/src/types/Product.ts
export type ProductType = 'phones' | 'tablets' | 'accessories';

export interface Product {
  id: number;
  name: string;
  fullPrice: number;
  price: number;
  image: string;
  category: string;
  capacity: string;
  screen: string;
  ram: string;
  itemId: string;
}

export interface CartProduct extends Product {
  quantity: number;
}

export interface ProductInfo
  extends Omit<Product, 'id' | 'image' | 'category' | 'itemId'> {
  id: string;
  namespace: string;
  color: string;
  category: ProductType;
  capacityAvailable: string[];
  colorsAvailable: string[];
  images: string[];
  descriptions: {
    title: string;
    text: string[];
  }[];
  resolution: string;
  processor: string;
  camera?: string;
  zoom?: string;
  cell: string[];
  itemId: number;
}